clean up city/county queries
the cities_counties table was not attaching the right city/county names to parcels--so i tried to fix it by putting the geom column at the end of the counties intersection, which seems to have made the spatial intersection work better. however, now it appears that the id being used to match for the parcels_in_cities table is not a zoning_id but instead a unique id of the row of the parcel_overlap table. to make the match in this query/table (parcel_in_cities) work we would want the id used to look up the city name from the zoning.codes_dictionary to be a zoning_id. unfortunately, the id used to match is the unique if of the parcel_overlap table, which does not have zoning_id's. so, must go back and run that query again. perhaps re-writing it. it seems silly to have to run these intersections so many times. it should just be 1 intersection.