add generic function for stats on overlapping geometries
this was necessary because each time we changed out a zoning geometry for a particular jurisdiction, we would have to re-create numerous tables, and re-run a parcel_overlaps table creation for everything that took more then 3 hours. now we can get statistics on multiple overlaps with parcels for a single jurisdiction, which can be used to add jurisdictions piecemeal to the zoning.parcel table.