add fixes to missing zoning codes and string matches
this adds zoning codes to the ID table that were not in that table, but which are existing fields in the source zoning data. it also adds string matching fixes to a number of cities. it needs a little work in the string matching piece to fill in the missing data. right now it only fills in/replaces 8 of 23 strings that need replacement--probably has to do with how postgres reads the strings from the csv or how the csv has them saved.