refactor plu '06 additions for 2-table data setup
we are doing this because its easier to edit zoning characteristics directly with a 2 table approach. the single table approach was done because we did not have plu '06 codes in the generic zoning table. however, this was easily resolved by adding them from the plu '06 source data. this generic zoning table now exists in bayarea_urbansim