only insert county zoning if zoning doesn't exist 4 parcel
otherwise, you end up with ~39k parcels that have multiple zoning assignments (or doubles of same assignment). this may be due to the simple single intersection at first, or city/county overlap. we do this by checking to see whether the parcel in the county zoning assignment already has a geom_id entry in the zoning.parcel table.