pass the zoning source name and the juris name through everywhere
currently, each time we decide which parcel belongs to which zoning polygon, we also look up the zoning_id from the zoning_lookup table by searching for the matching jurisdiction and zoning name (e.g. R-5-D). here we attempt to pass along the jurisdiction name and zoning name, in addition to the id that was looked up. the goal here is to be able to separate out that lookup (what the zoning_id is) from the answer to the question of which zoning geometry the parcel is mostly covered by. in the future, this will alow us just do the geometry processing separately from the editing of the zoning name attributes and characteristics for a parcel. then we could do some more of the zoning fixes and updates that are not about the geometry problem through python/pandas, which would be faster.