fix contested parcels ordering
we were selecting out contested parcels before assigning the new batch to them. this meant that contested parcels2 was effectively the same as contested_parcels, so we saw none of the benefits this selection was supposed to have on performance, and perhaps there may have been associated errors (though unlikely, because we only fill in geom_id's that have not been filled in with zoning_id's). same with contested_parcels3.