check (again) for missing zoning_id's in zoning_parcels.csv
there are some sql files in this repository (process/fill_in_missing_zoning_ids.sql), but they were used to modify the db piecemeal and not added back into the makefile. here we are adding a python script that checks for parcels in the zoning_parcels csv that have no generic zoning id, prints stats about those parcels, and if they are missing codes, looks up a code for them based on their jurisdictional data table name and the source zoning code assigned to them. in the future this allows us to work toward allowing non-postgres users to inspect, identify, and resolve issue related to zoning assignment that are purely related to the generic code dictionary (zoning_lookup.csv) and the source zoning names, as well as the assignment of a given generic zoning code to the csv.
i